LayoutTests/http/tests/css/prefers-color-scheme-in-nested-iframes-follows-system-preference-expected.txt

prefers-color-scheme in nested iframes follows system preference if none of the iframes explicitly specify color-scheme

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caughtEvent.data is "rgb(255, 0, 0)"
PASS caughtEvent.data is "rgb(255, 0, 0)"
PASS caughtEvent.data is "rgb(255, 0, 0)"
PASS caughtEvent.data is "rgb(255, 0, 0)"
PASS caughtEvent.data is "rgb(0, 255, 0)"
PASS caughtEvent.data is "rgb(0, 255, 0)"
PASS caughtEvent.data is "rgb(0, 255, 0)"
PASS caughtEvent.data is "rgb(0, 255, 0)"
PASS successfullyParsed is true

TEST COMPLETE
